fix(review): 리뷰 별점 0.5단위로 수정

+++ services/review-service/src/main/java/com/golajugaenyang/review/adapter/in/web/dto/response/ReviewDetailResponse.java
@@ -8,7 +8,7 @@ public record ReviewDetailResponse(
         boolean isMine,
         Product product,
         Long petId,
-        int rating,
+        double rating,
         int usagePeriod,
         List<AnswerValue> answerValues,
         List<String> goodPoints,
