+++ services/review-service/src/main/java/com/golajugaenyang/review/adapter/in/web/dto/response/ReviewFilterListResponse.java
@@ -12,7 +12,7 @@ public record ReviewFilterListResponse(
             Long reviewId,
             String nickname,
             Pet pet,
-            int rating,
+            double rating,
             String usagePeriod,
             String palatability,
             String text,
